RequireAuth with requireSuperUser & redirect props
